Minor grammatical modifications [1 to ~1000] - FR
It only concerns the first thousand lines. The original is already very qualitative so there isn't much to modify. I only did a couple changes on what I consider weirdly phrased sentences so that there's better grammar overall. I'll do the rest later, which means I'll be proposing about 6 changes (including this one), less if I modify more of the code in the future.
Ps : To anyone reading this, have a good day or evening.